modified course_projekt

+++ course_project/library/ground_transport.h
@@ -11,13 +11,13 @@
 class GroundTransport : public Transport {
 protected:
     int movementTime_;
-    int restTime_;
+    double restTime_;
 public:
     LIBRARY_API GroundTransport();
     LIBRARY_API GroundTransport(int s);
 
     LIBRARY_API int getMovementTime();
-    LIBRARY_API int getRestTime(int s);
+    LIBRARY_API double getRestTime(int s);
 
     LIBRARY_API int getStopNumber(int s);
     LIBRARY_API double getTotalTime(int s) override;
